+++ mobile/android/android-components/components/service/glean/src/main/java/mozilla/components/service/glean/private/MetricAliases.kt
@@ -1,0 +1,28 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+package mozilla.components.service.glean.private
+
+typealias Lifetime = mozilla.telemetry.glean.private.Lifetime
+typealias NoExtraKeys = mozilla.telemetry.glean.private.NoExtraKeys
+
+typealias BooleanMetricType = mozilla.telemetry.glean.private.BooleanMetricType
+typealias CounterMetricType = mozilla.telemetry.glean.private.CounterMetricType
+typealias CustomDistributionMetricType = mozilla.telemetry.glean.private.CustomDistributionMetricType
+typealias DatetimeMetricType = mozilla.telemetry.glean.private.DatetimeMetricType
+typealias EventMetricType<T> = mozilla.telemetry.glean.private.EventMetricType<T>
+typealias HistogramMetricBase = mozilla.telemetry.glean.private.HistogramBase
+typealias HistogramType = mozilla.telemetry.glean.private.HistogramType
+typealias LabeledMetricType<T> = mozilla.telemetry.glean.private.LabeledMetricType<T>
+typealias MemoryDistributionMetricType = mozilla.telemetry.glean.private.MemoryDistributionMetricType
+typealias MemoryUnit = mozilla.telemetry.glean.private.MemoryUnit
+typealias PingType = mozilla.telemetry.glean.private.PingType
+typealias QuantityMetricType = mozilla.telemetry.glean.private.QuantityMetricType
+typealias RecordedExperimentData = mozilla.telemetry.glean.private.RecordedExperimentData
+typealias StringListMetricType = mozilla.telemetry.glean.private.StringListMetricType
+typealias StringMetricType = mozilla.telemetry.glean.private.StringMetricType
+typealias TimespanMetricType = mozilla.telemetry.glean.private.TimespanMetricType
+typealias TimeUnit = mozilla.telemetry.glean.private.TimeUnit
+typealias TimingDistributionMetricType = mozilla.telemetry.glean.private.TimingDistributionMetricType
+typealias UuidMetricType = mozilla.telemetry.glean.private.UuidMetricType